move tracing setup to its own onboarding page
