+++ web/src/features/setup/components/SetupPage.tsx
@@ -10,67 +10,36 @@ import {
 import { Button } from "@/src/components/ui/button";
 import { Card } from "@/src/components/ui/card";
 import { NewOrganizationForm } from "@/src/features/organizations/components/NewOrganizationForm";
-import { usePostHogClientCapture } from "@/src/features/posthog-analytics/usePostHogClientCapture";
 import { NewProjectForm } from "@/src/features/projects/components/NewProjectForm";
 import { useQueryProjectOrOrganization } from "@/src/features/projects/hooks";
-import { ApiKeyRender } from "@/src/features/public-api/components/CreateApiKeyButton";
-import { QuickstartExamples } from "@/src/features/public-api/components/QuickstartExamples";
 import { MembershipInvitesPage } from "@/src/features/rbac/components/MembershipInvitesPage";
 import { MembersTable } from "@/src/features/rbac/components/MembersTable";
 import {
   createProjectRoute,
   inviteMembersRoute,
-  setupTracingRoute,
 } from "@/src/features/setup/setupRoutes";
-import { api } from "@/src/utils/api";
 import { cn } from "@/src/utils/tailwind";
-import { type RouterOutput } from "@/src/utils/types";
 import { Check } from "lucide-react";
 import { useRouter } from "next/router";
-import { useEffect, useState } from "react";
 import { StringParam, useQueryParam } from "use-query-params";
 
 // Multi-step setup process
 // 1. Create Organization: /setup
 // 2. Invite Members: /organization/:orgId/setup
 // 3. Create Project: /organization/:orgId/setup?step=create-project
-// 4. Setup Tracing: /project/:projectId/setup
 export function SetupPage() {
   const { project, organization } = useQueryProjectOrOrganization();
   const router = useRouter();
   const [orgStep] = useQueryParam("orgstep", StringParam); // "invite-members" | "create-project"
-  const queryProjectId = router.query.projectId as string | undefined;
 
   // starts at 1 to align with breadcrumb
   const stepInt = !organization
     ? 1
     : project
-      ? 4
+      ? 3
       : orgStep === "create-project"
         ? 3
         : 2;
-
-  const hasTracingConfigured = api.traces.hasTracingConfigured.useQuery(
-    { projectId: queryProjectId as string },
-    {
-      enabled: queryProjectId !== undefined && stepInt === 4,
-      refetchInterval: 5000,
-      trpc: {
-        context: {
-          skipBatch: true,
-        },
-      },
-    },
-  ).data;
-
-  const capture = usePostHogClientCapture();
-  useEffect(() => {
-    if (hasTracingConfigured !== undefined) {
-      capture("onboarding:tracing_check_active", {
-        active: hasTracingConfigured,
-      });
-    }
-  }, [hasTracingConfigured, capture]);
 
   return (
     <ContainerPage
@@ -127,20 +96,6 @@ export function SetupPage() {
               )}
             >
               3. Create Project
-              {stepInt > 3 && <Check className="ml-1 inline-block h-3 w-3" />}
-            </BreadcrumbPage>
-          </BreadcrumbItem>
-          <BreadcrumbSeparator />
-          <BreadcrumbItem>
-            <BreadcrumbPage
-              className={cn(
-                stepInt !== 4
-                  ? "text-muted-foreground"
-                  : "font-semibold text-foreground",
-              )}
-            >
-              4. Setup Tracing
-              {stepInt === 4 && <Check className="ml-1 inline-block h-3 w-3" />}
             </BreadcrumbPage>
           </BreadcrumbItem>
         </BreadcrumbList>
@@ -193,19 +148,10 @@ export function SetupPage() {
               <NewProjectForm
                 orgId={organization.id}
                 onSuccess={(projectId) =>
-                  router.push(setupTracingRoute(projectId))
+                  router.push(`/project/${projectId}/traces`)
                 }
               />
             </div>
-          )
-        }
-        {
-          // 4. Setup Tracing
-          stepInt === 4 && project && organization && (
-            <TracingSetup
-              projectId={project.id}
-              hasTracingConfigured={hasTracingConfigured ?? false}
-            />
           )
         }
       </Card>
@@ -219,88 +165,6 @@ export function SetupPage() {
           Next
         </Button>
       )}
-      {
-        // 4. Setup Tracing
-        stepInt === 4 && project && (
-          <Button
-            className="mt-4 self-start"
-            onClick={() => router.push(`/project/${project.id}`)}
-            variant={hasTracingConfigured ? "default" : "secondary"}
-          >
-            {hasTracingConfigured ? "Open Dashboard" : "Skip for now"}
-          </Button>
-        )
-      }
     </ContainerPage>
   );
 }
-
-const TracingSetup = ({
-  projectId,
-  hasTracingConfigured,
-}: {
-  projectId: string;
-  hasTracingConfigured?: boolean;
-}) => {
-  const [apiKeys, setApiKeys] = useState<
-    RouterOutput["projectApiKeys"]["create"] | null
-  >(null);
-  const utils = api.useUtils();
-  const mutCreateApiKey = api.projectApiKeys.create.useMutation({
-    onSuccess: (data) => {
-      utils.projectApiKeys.invalidate();
-      setApiKeys(data);
-    },
-  });
-
-  const createApiKey = async () => {
-    try {
-      await mutCreateApiKey.mutateAsync({ projectId });
-    } catch (error) {
-      console.error("Error creating API key:", error);
-    }
-  };
-
-  return (
-    <div className="space-y-8">
-      <div>
-        <Header title="API Keys" />
-        <p className="mb-4 text-sm text-muted-foreground">
-          These keys are used to authenticate your API requests. You can create
-          more keys later in the project settings.
-        </p>
-        {apiKeys ? (
-          <ApiKeyRender generatedKeys={apiKeys} scope={"project"} />
-        ) : (
-          <div className="flex flex-col gap-4">
-            <p className="text-sm text-muted-foreground">
-              You need to create an API key to start tracing your application.
-            </p>
-            <Button
-              onClick={createApiKey}
-              loading={mutCreateApiKey.isPending}
-              className="self-start"
-            >
-              Create API Key
-            </Button>
-          </div>
-        )}
-      </div>
-
-      <div>
-        <Header
-          title="Setup Tracing"
-          status={hasTracingConfigured ? "active" : "pending"}
-        />
-        <p className="mb-4 text-sm text-muted-foreground">
-          Tracing is used to track and analyze your LLM calls. You can always
-          skip this step and setup tracing later.
-        </p>
-        <QuickstartExamples
-          secretKey={apiKeys?.secretKey}
-          publicKey={apiKeys?.publicKey}
-        />
-      </div>
-    </div>
-  );
-};
